+++ backend/src/main/java/ch/zhaw/parkship/tag/TagDto.java
@@ -2,7 +2,6 @@ package ch.zhaw.parkship.tag;
 
 import java.io.Serializable;
 import java.util.Objects;
-
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
@@ -10,40 +9,40 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TagDto implements Serializable {
-    private Integer id;
+  private Long id;
 
-    @NotBlank
-    private String name;
+  @NotBlank
+  private String name;
 
-    @NotBlank
-    private String category;
+  @NotBlank
+  private String category;
 
-    public TagDto(TagEntity tagEntity) {
-        this.id = tagEntity.getId();
-        this.name = tagEntity.getName();
-        this.category = tagEntity.getCategory();
-    }
+  public TagDto(TagEntity tagEntity) {
+    this.id = tagEntity.getId();
+    this.name = tagEntity.getName();
+    this.category = tagEntity.getCategory();
+  }
 
-    public TagDto() {
-    }
+  public TagDto() {}
 
-    @Override
-    public String toString() {
-        return "TagDto{" + "id=" + id + ", name='" + name + '\'' + ", category='" + category + '\'' + '}';
-    }
+  @Override
+  public String toString() {
+    return "TagDto{" + "id=" + id + ", name='" + name + '\'' + ", category='" + category + '\''
+        + '}';
+  }
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
+  @Override
+  public int hashCode() {
+    return Objects.hash(id);
+  }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (!(o instanceof TagDto))
-            return false;
-        TagDto tagDto = (TagDto) o;
-        return Objects.equals(id, tagDto.id);
-    }
+  @Override
+  public boolean equals(Object o) {
+    if (this == o)
+      return true;
+    if (!(o instanceof TagDto))
+      return false;
+    TagDto tagDto = (TagDto) o;
+    return Objects.equals(id, tagDto.id);
+  }
 }
