+++ backend/src/main/java/ch/zhaw/parkship/tag/TagEntity.java
@@ -3,7 +3,6 @@ package ch.zhaw.parkship.tag;
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
-
 import ch.zhaw.parkship.parkinglot.ParkingLotEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -22,41 +21,41 @@ import lombok.Setter;
 @Table(name = "Tags")
 public class TagEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
 
-    @Column(nullable = false)
-    private String name;
+  @Column(nullable = false)
+  private String name;
 
-    @Column(nullable = false)
-    private String category;
+  @Column(nullable = false)
+  private String category;
 
-    @ManyToMany(mappedBy = "tags", fetch = FetchType.EAGER)
-    private Set<ParkingLotEntity> parkingLots;
+  @ManyToMany(mappedBy = "tags", fetch = FetchType.EAGER)
+  private Set<ParkingLotEntity> parkingLots;
 
-    public TagEntity() {
-        this.parkingLots = new HashSet<>();
-    }
+  public TagEntity() {
+    this.parkingLots = new HashSet<>();
+  }
 
-    @Override
-    public String toString() {
-        return "TagEntity{" + "id=" + id + ", name='" + name + '\'' + ", category='" + category + '\''
-                + ", parkingLots=" + parkingLots + '}';
-    }
+  @Override
+  public String toString() {
+    return "TagEntity{" + "id=" + id + ", name='" + name + '\'' + ", category='" + category + '\''
+        + ", parkingLots=" + parkingLots + '}';
+  }
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
+  @Override
+  public int hashCode() {
+    return Objects.hash(id);
+  }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (!(o instanceof TagEntity))
-            return false;
-        TagEntity tag = (TagEntity) o;
-        return Objects.equals(id, tag.id);
-    }
+  @Override
+  public boolean equals(Object o) {
+    if (this == o)
+      return true;
+    if (!(o instanceof TagEntity))
+      return false;
+    TagEntity tag = (TagEntity) o;
+    return Objects.equals(id, tag.id);
+  }
 }
